Ignore bugs created for intermittent failures in build and test automation for the following queries: - total new - total closed - median time to fix a bug - open bugs with bugzilla severity 'normal' or 'minor'
There are 200+ bugs created each week for intermittent failures. A few are
frequent or other important, others fail once and likely an issue with the test
(most often a race condition). An import of tests like web-platform-tests (wpt)
can increase that level and distort the bug counts but it doesn't affect quality
of the unchanged code which ships as product.

For failures detected by build and test automation, use these queries:
- New: CI Automation: bugs/100 pushes
- New: CI Automation: crash bugs/100 pushes